+++ test/server/jest.config.js
@@ -1,0 +1,13 @@
+const config = require('./config.json');
+
+const jestConfig = {
+  verbose: true,
+  globals: {
+    ...config
+  }
+};
+
+module.exports = () => {
+  return jestConfig;
+};
+