+++ test/server/services/simple.options.js
@@ -1,0 +1,2 @@
+_out.json({result: true})
+